added support for app going into background and resuming the game in the paused state (appdelegate,lvl1,nsnotificationcenter),fixed player max movement in y direction, silenced type conversion errors in JSTileMap, fixed memory leak in lvl2, reduced code duplication in lvl1 related to button highlighting, and created door class

+++ Metroidvania/JSTileMap.m
@@ -117,15 +117,15 @@
         return;
     }
     
-    TMXTilesetInfo* tilesetInfo = [self.map tilesetInfoForGid:gID];
+    TMXTilesetInfo* tilesetInfo = [self.map tilesetInfoForGid:(int)gID];
     [self.tileInfo addObject:tilesetInfo];
     
     SKSpriteNode *tile = [self tileAtCoord:coord];
     if (tilesetInfo && tile){
-        SKTexture* texture = [tilesetInfo textureForGid:gID];
+        SKTexture* texture = [tilesetInfo textureForGid:(int)gID];
         if (texture){
             [tile setTexture:texture];
-            _layerInfo.tiles[ idx ] = gID;
+            _layerInfo.tiles[ idx ] = (int)gID;
         }
     }
 }
